requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.13.1
fonttools==4.58.4
fsspec==2024.6.1
huggingface-hub==0.32.4
idna==3.10
intel-openmp==2021.4.0
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.1
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
pip==25.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.16.1
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
tbb==2021.13.1
torch==2.3.1
torchaudio==2.5.1+cu121
torchvision==0.18.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.4.0
vggt==0.0.1